+++ wombat/core/environment.py
@@ -2,19 +2,19 @@
 from __future__ import annotations
 
 import csv
-import math  # type: ignore
-import logging  # type: ignore
-import datetime as dt  # type: ignore
-from typing import TYPE_CHECKING, Tuple, Union, Optional  # type: ignore
-from pathlib import Path  # type: ignore
-from datetime import datetime, timedelta  # type: ignore
+import math
+import logging
+import datetime as dt
+from typing import TYPE_CHECKING, Tuple, Union, Optional
+from pathlib import Path
+from datetime import datetime, timedelta
 
-import numpy as np  # type: ignore
-import simpy  # type: ignore
-import pandas as pd  # type: ignore
+import numpy as np
+import simpy
+import pandas as pd
 import pyarrow as pa
 import pyarrow.csv  # pylint: disable=W0611
-from simpy.events import Event  # type: ignore
+from simpy.events import Event
 from pandas.core.indexes.datetimes import DatetimeIndex
 
 import wombat  # pylint: disable=W0611
@@ -41,6 +41,7 @@ EVENTS_COLUMNS = [
     "system_operating_level",
     "part_operating_level",
     "duration",
+    "distance_km",
     "request_id",
     "location",
     "materials_cost",
@@ -148,10 +149,10 @@ class WombatEnvironment(simpy.Environment):
 
         self.workday_start = int(workday_start)
         self.workday_end = int(workday_end)
-        if not 0 <= self.workday_start < 24:
-            raise ValueError("workday_start must be a valid 24hr time before midnight!")
-        if not 0 <= self.workday_end < 24:
-            raise ValueError("workday_end must be a valid 24hr time before midnight!")
+        if not 0 <= self.workday_start <= 24:
+            raise ValueError("workday_start must be a valid 24hr time before midnight.")
+        if not 0 <= self.workday_end <= 24:
+            raise ValueError("workday_end must be a valid 24hr time.")
         if self.workday_end <= self.workday_start:
             raise ValueError(
                 "Work shifts must end after they start ({self.workday_start}hrs)."
@@ -294,7 +295,7 @@ class WombatEnvironment(simpy.Environment):
             True if it's valid working hours, False otherwise.
         """
         if -1 in (workday_start, workday_end):
-            # Return true if the shift is around the clock
+            # Return True if the shift is around the clock
             if self.workday_start == 0 and self.workday_end == 24:
                 return True
             return self.workday_start <= self.simulation_time.hour < self.workday_end
@@ -535,6 +536,7 @@ class WombatEnvironment(simpy.Environment):
         system_ol: float | int = 0,
         part_ol: float | int = 0,
         duration: float = 0,
+        distance_km: float = 0,
         request_id: str = "na",
         location: str = "na",
         materials_cost: Union[int, float] = 0,
@@ -574,6 +576,8 @@ class WombatEnvironment(simpy.Environment):
             enroute, or system, by default "na".
         duration : float
             Length of time the action lasted, by default 0.
+        distance : float
+            Distance traveled, in km, if applicable, by default 0.
         materials_cost : Union[int, float], optional
             Total cost of materials for action, in USD, by default 0.
         hourly_labor_cost : Union[int, float], optional
@@ -605,6 +609,7 @@ class WombatEnvironment(simpy.Environment):
             reason=reason,
             additional=additional,
             duration=duration,
+            distance_km=distance_km,
             request_id=request_id,
             location=location,
             materials_cost=materials_cost,